+++ database/migrations/2022_03_22_082508_create_pots_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
     {
         Schema::create('pots', function (Blueprint $table) {
             $table->increments('id');
-            $table->float('amount');
+            $table->decimal('amount', 14, 2);
             $table->foreignIdFor(Hand::class);
             $table->timestamps();
         });
